5072 - Add RemoteSensingSurvey video link
