+++ src/client/pages/Kiosk/components/Video/Video.scss
@@ -1,17 +1,17 @@
 @use 'sass:color';
 
-.kiosk-fra-process {
+.kiosk-video {
   height: 100%;
   position: relative;
   width: 100%;
 }
 
-.kiosk-fra-process__yt-container {
+.kiosk-video__player {
   height: 100%;
   width: 100%;
 }
 
-.kiosk-fra-process__yt-video-overlay {
+.kiosk-video__overlay {
   align-items: center;
   background-color: rgba(0, 0, 0, 0.9);
   bottom: 0;
@@ -24,8 +24,7 @@
   z-index: 1;
 }
 
-.kiosk-fra-process__yt-video-overlay-button {
-  background-color: #61a2d8;
+.kiosk-video__overlay-button {
   border-radius: 30px;
   display: flex;
   font-size: min(1.78dvw, 3dvh);
@@ -33,6 +32,7 @@
   line-height: unset;
   min-height: 4dvh;
   padding: min(1.8dvw, 2.55dvh);
+  background-color: #61a2d8;
 
   &:hover {
     background-color: color.adjust(#61a2d8, $lightness: 5%, $space: hsl);
