+++ src/client/pages/Kiosk/components/Video/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Video'
